Refactor response library
Create map of default messages to be returned for a given status code.
Also implement NewErrorResponse method and use it for generating all
ErrorResponses.

Create methods for common response statuses with a default message.  Also
accept zero or more strings on most Err methods.

This allows for the convenient and consistent generation and testing of
error responses.  Also, updating existing defaults should be passed through
to all tests automatically.